+++ ChangeLog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Null pointer dereference exception on `accessorData` (thanks [@hybridherbst](https://github.com/hybridherbst))
+
 ## [4.0.0] - 2021-05-21
 ### Added
 - Import glTF files at design-time in the Editor